# Lower memset and memcpy to 8051 for performance