The performance of many real world applications is dominated by data movement, and this provides an opportunity to speed rv51 by implementing a few key libc functions at the 8051 level. At the cost of an extra 166 bytes (minus a small improvement in RISCV size) I found some nice speedups:

## RISCV benchmarks

| Benchmark | memcpy/set change | rv51 (s51 ticks) | Speedup |
|--------- |----------------- |---------------- |------- |
| median    | 20799744          | 28880352         | 28.0%   |
| qsort     | 450237564         | 488045772        | 7.7%    |
| rsort     | 834642576         | 884368992        | 5.6%    |
| towers    | 17660928          | 17799816         | 0.8%    |
| vvadd     | 12810540          | 21687048         | 40.9%   |
| multiply  | 83308452          | 86778900         | 4.0%    |


## Applications

| Name            | with new memcpy/memset | without       | Speedup |
|--------------- |---------------------- |------------- |------- |
| C hello world   | 3259980 ticks          | 3324624 ticks | 1.9%    |
| My TinyUSB fork | 13s                    | 20s           | 35.0%   |

## Architecture

I wasn't sure how to properly handle the transition between layers. For now it's a system call. If there is a designated mechanism in RISCV for communicating between code and e.g. a virtual machine that would be a better fit.

I hope this will be interesting upstream; I know I intend to use it in my TinyUSB project where it's very helpful in filling the gap between the speed of sdcc (such as it is) and rv51.