@jefftrull 2024-03-27
Maybe this fits in part of the SBI space somewhere https://drive.google.com/file/d/1U2kwjqxXgDONXk_-ZDTYzvsV-F_8ylEH/view