@cyrozap 2024-06-18
While this is really interesting and seems useful, and it's clear you've put a lot of effort into it (thank you for including the benchmarks), I'm not going to include these changes in my fork of rv51. While I did explicitly design rv51 to be extensible (e.g., to add accelerator functions like this one or to interface with data movement/processing peripherals), I would really like to keep the code in this repository as minimal as possible. That is, my intention is to support the RISC-V _ISA_, not particular functions from specific programming languages (even though memcpy is pretty much universal). Thank you.